backend/requirements.txt
fastapi>=0.100.0,<1.0
groq
langchain-groq
google-generativeai
langchain-google-genai
pydantic[email]>=1.10.4
requests>=2.22.0
loguru>=0.6.0
joblib>=1.2.0
certifi>=2022.12.7
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiohttp>=3.8.3
fastapi-utils>=0.2.1
envyaml>=1.10.211231
tokenizers>=0.13.2
pydub>=0.25.1
llama_index>=0.9.10
nest_asyncio
httpx
sqlalchemy
databases
sqlite-utils
qdrant-client
nltk>=3.8.1
PyMuPDF>=1.23.22
llama-index-vector-stores-qdrant
uvicorn
pandas
